feat: add SaveListForm component with form validation for adding and editing lists
