+++ package.json
@@ -29,6 +29,7 @@
     "react-router": "^7.0.2",
     "tailwind-merge": "^2.5.5",
     "tailwindcss-animate": "^1.0.7",
+    "usehooks-ts": "^3.1.0",
     "uuid": "^11.0.3",
     "vaul": "^1.1.2",
     "zod": "^3.24.1",
